Save of tracks, albums, and artists to DynamoDB.
