+++ src/main/java/f18a14c09s/integration/alexa/music/catalog/data/StringPkIntegerSkDynamoDbItem.java
@@ -1,0 +1,11 @@
+package f18a14c09s.integration.alexa.music.catalog.data;
+
+public interface StringPkIntegerSkDynamoDbItem {
+    String getPk();
+
+    Integer getSk();
+
+    void setPk(String pk);
+
+    void setSk(Integer sk);
+}
