+++ src/main/java/f18a14c09s/integration/alexa/music/catalog/data/StringPkStringSkDynamoDbItem.java
@@ -1,0 +1,11 @@
+package f18a14c09s.integration.alexa.music.catalog.data;
+
+public interface StringPkStringSkDynamoDbItem {
+    String getPk();
+
+    String getSk();
+
+    void setPk(String pk);
+
+    void setSk(String sk);
+}
